+++ script/SafeConfig.sol
@@ -6,9 +6,7 @@ import {ItoA} from "src/utils/ItoA.sol";
 library SafeConfig {
     function _isTestnet() internal view returns (bool) {
         if (
-            block.chainid == 10143 // monad testnet
-                || block.chainid == 11124 // abstract sepolia
-                || block.chainid == 11155111 // sepolia
+            block.chainid == 11155111 // sepolia
         ) {
             return true;
         }
@@ -59,8 +57,6 @@ library SafeConfig {
                 || block.chainid == 5000 // mantle
                 || block.chainid == 8453 // base
                 || block.chainid == 9745 // plasma
-                || block.chainid == 10143 // monad testnet
-                || block.chainid == 11124 // abstract sepolia
                 || block.chainid == 34443 // mode
                 || block.chainid == 42161 // arbitrum
                 || block.chainid == 43114 // avalanche
@@ -79,7 +75,6 @@ library SafeConfig {
     function isEraVm() internal view returns (bool) {
         if (
             block.chainid == 2741 // abstract
-                || block.chainid == 11124 // abstract sepolia
         ) {
             return true;
         }
@@ -89,6 +84,7 @@ library SafeConfig {
                 || block.chainid == 56 // bnb
                 || block.chainid == 130 // unichain
                 || block.chainid == 137 // polygon
+                || block.chainid == 143 // monad
                 || block.chainid == 146 // sonic
                 || block.chainid == 480 // worldchain
                 || block.chainid == 999 // hyperevm
@@ -96,7 +92,6 @@ library SafeConfig {
                 || block.chainid == 5000 // mantle
                 || block.chainid == 8453 // base
                 || block.chainid == 9745 // plasma
-                || block.chainid == 10143 // monad testnet
                 || block.chainid == 34443 // mode
                 || block.chainid == 42161 // arbitrum
                 || block.chainid == 43114 // avalanche