+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * DackieSwap and BaseX UniV3 forks removed from Base
 * Remove `LFJTM` action for Let's F***ing Joe Token Mill V2 on Monad mainnet chain
 * Abandon the following chains:
+  * Monad testnet
+  * Abstract Sepolia
   * Blast
   * Taiko
   * Gnosis (XDAI)